Deduplicate some common functionality
