+++ imessage-exporter/src/exporters/mod.rs
@@ -1,3 +1,4 @@
 pub mod exporter;
 pub mod html;
+mod shared;
 pub mod txt;
