+++ imessage-exporter/src/exporters/txt.rs
@@ -5,7 +5,7 @@ use std::{
     },
     fmt::Write as FmtWrite,
     fs::File,
-    io::{BufWriter, Write},
+    io::BufWriter,
 };
 
 use crate::{
@@ -24,7 +24,6 @@ use imessage_database::{
         collaboration::CollaborationMessage,
         digital_touch::{self, DigitalTouch},
         edited::{EditStatus, EditedMessage},
-        expressives::{BubbleEffect, Expressive, ScreenEffect},
         handwriting::HandwrittenMessage,
         music::MusicMessage,
         placemark::PlacemarkMessage,
@@ -162,11 +161,6 @@ impl<'a> Exporter<'a> for TXT<'a> {
             }
             None => Ok(&mut self.orphaned),
         }
-    }
-
-    fn write_to_file(file: &mut BufWriter<File>, text: &str) -> Result<(), RuntimeError> {
-        file.write_all(text.as_bytes())
-            .map_err(RuntimeError::DiskError)
     }
 }
 
@@ -617,27 +611,7 @@ impl<'a> MessageFormatter<'a> for TXT<'a> {
     }
 
     fn format_expressive(&self, msg: &'a Message) -> &'a str {
-        match msg.get_expressive() {
-            Expressive::Screen(effect) => match effect {
-                ScreenEffect::Confetti => "Sent with Confetti",
-                ScreenEffect::Echo => "Sent with Echo",
-                ScreenEffect::Fireworks => "Sent with Fireworks",
-                ScreenEffect::Balloons => "Sent with Balloons",
-                ScreenEffect::Heart => "Sent with Heart",
-                ScreenEffect::Lasers => "Sent with Lasers",
-                ScreenEffect::ShootingStar => "Sent with Shooting Star",
-                ScreenEffect::Sparkles => "Sent with Sparkles",
-                ScreenEffect::Spotlight => "Sent with Spotlight",
-            },
-            Expressive::Bubble(effect) => match effect {
-                BubbleEffect::Slam => "Sent with Slam",
-                BubbleEffect::Loud => "Sent with Loud",
-                BubbleEffect::Gentle => "Sent with Gentle",
-                BubbleEffect::InvisibleInk => "Sent with Invisible Ink",
-            },
-            Expressive::Unknown(effect) => effect,
-            Expressive::None => "",
-        }
+        super::shared::format_expressive(msg)
     }
 
     fn format_announcement(&self, msg: &'a Message) -> String {
@@ -1170,19 +1144,12 @@ impl<'a> BalloonFormatter<&'a str> for TXT<'a> {
 // MARK: Impl
 impl TXT<'_> {
     fn get_time(&self, message: &Message) -> String {
-        let mut date = format(&message.date(&self.config.offset));
-        let read_after = message.time_until_read(&self.config.offset);
-        if let Some(time) = read_after
-            && !time.is_empty()
-        {
-            let who = if message.is_from_me() {
-                "them"
-            } else {
-                self.config.options.custom_name.as_deref().unwrap_or("you")
-            };
-            let _ = write!(date, " (Read by {who} after {time})");
+        let (date, read_receipt) = super::shared::message_time(self.config, message);
+        if read_receipt.is_empty() {
+            date
+        } else {
+            format!("{date} {read_receipt}")
         }
-        date
     }
 
     fn add_line(&self, string: &mut String, part: &str, indent: &str) {
